Contains work on selection using a "hack" of changing colors of objects unseen to users and using the new colors as IDs for the objects.
Problems: This currenty does not work with the new MinVR config files for
the Fake Head Tracker. There appears to be a scaling issue where, when
running the program, the camera is either pointed in the wrong direction
or the scene has the wrong scale. This does work on a computer screen, with
a mouse, with the old config file.

To Run: bin/selection ../shaders/shader2.vp ../shaders/shader.fp -c ../config/desktop-freeglut.xml